+++ dzVents/runtime/misc/testdzVents.sh
@@ -142,7 +142,7 @@ function fillNumberOfTests
 		Lodash_ExpectedTests=100
 		ScriptdzVentsDispatching_ExpectedTests=2
 		TimedCommand_ExpectedTests=46
-		Time_ExpectedTests=342
+		Time_ExpectedTests=357
 		Utils_ExpectedTests=27
 		Variable_ExpectedTests=15
 		ContactDoorLockInvertedSwitch_ExpectedTests=2
@@ -267,6 +267,6 @@ else
 fi
 
 echo Total tests: $totalTest
-echo $(date)';' dzVents version $dzVersion tested without erors after $(showTime)
+echo $(date)';' dzVents version $dzVersion tested without errors after $(showTime)
 stopBackgroundProcesses 0
 cleanup