+++ dzVents/documentation/history.md
@@ -1,3 +1,7 @@
+[3.0.11]
+- Add sensorValue attribute to custom sensor
+- Add sunatsouth as moment in time (like sunrise / sunset ) 
+
 [3.0.10]
 - Add NSS_GOOGLE_DEVICES for notification casting to Google home / Google chromecast
 - Add optional parm delay to domoticz.sendCommand, domoticz.email, domoticz.sms and domoticz.notify